Fix bug Tax amount = 0
Convert int TAX = 8 to decimal TAX = 8

# Conflicts:
#	.vs/SupermarketAPI/FileContentIndex/b1db3c56-7f44-4993-a49f-5c831e940b97.vsidx
#	.vs/SupermarketAPI/v17/.futdcache.v2
#	.vs/SupermarketAPI/v17/.suo
#	.vs/SupermarketAPI/v17/DocumentLayout.backup.json
#	.vs/SupermarketAPI/v17/DocumentLayout.json
#	bin/Debug/net8.0/SupermarketAPI.dll
#	bin/Debug/net8.0/SupermarketAPI.exe
#	bin/Debug/net8.0/SupermarketAPI.pdb
#	obj/Debug/net8.0/SupermarketAPI.AssemblyInfo.cs
#	obj/Debug/net8.0/SupermarketAPI.AssemblyInfoInputs.cache
#	obj/Debug/net8.0/SupermarketAPI.dll
#	obj/Debug/net8.0/SupermarketAPI.pdb
#	obj/Debug/net8.0/SupermarketAPI.sourcelink.json
#	obj/Debug/net8.0/apphost.exe
#	obj/Debug/net8.0/ref/SupermarketAPI.dll
#	obj/Debug/net8.0/refint/SupermarketAPI.dll

+++ Services/Impl/OrderService.cs
@@ -11,7 +11,7 @@ namespace SupermarketAPI.Services.Impl
     {
         private readonly IOrderRepository _orderRepository;
         private readonly IBillRepository _billRepository;
-        private int TAX_PERCENT = 8;
+        private decimal TAX_PERCENT = 8;
 
         public OrderService(IOrderRepository orderRepository, IBillRepository billRepository)
         {
